refactor(setting): generalize setting type to use any type parameter
Update the Setting type to be a generic function type, allowing it to accept
any type parameter S. This change simplifies the Apply function signature by
removing the redundant type parameter and enhances the flexibility of the
setting package.

BREAKING CHANGE: This change may affect consumers of the setting package who
rely on the previous non-generic definition of the Setting type.